Safari-609.1.11.1/LayoutTests/js/dom/modules/module-async-and-window-load-expected.txt

Test window.load is fired after module is executed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Module is not executed yet.
PASS moduleIsExecuted is true
PASS successfullyParsed is true

TEST COMPLETE
